+++ exonum/src/lib.rs
@@ -17,6 +17,7 @@
 //! For more information see the project readme.
 
 #![deny(missing_debug_implementations)]
+#![deny(missing_docs)]
 
 
 #![cfg_attr(feature="cargo-clippy", allow(zero_prefixed_literal))]